Stubbed out update for enrichment via cron
