+++ src/server/routes/enrichment/visualization/update-enrichment.js
@@ -1,0 +1,3 @@
+const updateEnrichment = () => {};
+
+module.exports = { updateEnrichment };
