+++ src/server/update.js
@@ -1,0 +1,7 @@
+const { updateEnrichment } = require('./routes/enrichment/visualization');
+
+const update = () => {
+  updateEnrichment();
+};
+
+ module.exports = update;
